+++ R/jaatha.R
@@ -1,11 +1,3 @@
-#' Jaatha: Simulation-based maximum likelihood parameter estimation
-#' 
-#' Jaatha is a simulation based maximum likelihood parameter estimation
-#' methods.
-#' 
-#' @name jaatha-package
-#' @docType package
-#' @keywords package
 #' @importFrom parallel mclapply
 #' @importFrom assertthat assert_that is.string is.count
 NULL