+++ NEWS.md
@@ -15,7 +15,7 @@ jaatha 3.1.0 (in development)
 * Increases robustsness to not-converging GLMs. If a GLM fails to converge, now
   only the current repetition is aborded instead of the compelete analysis 
   (#110, #111). 
-* Dump frames to tempdir() on simulation errors.
+* Dump frames to tempdir() on simulation errors (#113).